+++ scripts/pylib/pytest-twister-harness/src/twister_harness/__init__.py
@@ -5,7 +5,7 @@
 # flake8: noqa
 
 from twister_harness.device.device_adapter import DeviceAdapter
-from twister_harness.fixtures.mcumgr import MCUmgr
+from twister_harness.helpers.mcumgr import MCUmgr
 from twister_harness.helpers.shell import Shell
 
 __all__ = ['DeviceAdapter', 'MCUmgr', 'Shell']
