+++ scripts/pylib/pytest-twister-harness/src/twister_harness/fixtures.py
@@ -10,6 +10,8 @@ import pytest
 from twister_harness.device.device_adapter import DeviceAdapter
 from twister_harness.device.factory import DeviceFactory
 from twister_harness.twister_harness_config import DeviceConfig, TwisterHarnessConfig
+from twister_harness.helpers.shell import Shell
+from twister_harness.helpers.mcumgr import MCUmgr
 
 logger = logging.getLogger(__name__)
 
@@ -44,3 +46,23 @@ def dut(request: pytest.FixtureRequest, device_object: DeviceAdapter) -> Generat
         yield device_object
     finally:  # to make sure we close all running processes execution
         device_object.close()
+
+
+@pytest.fixture(scope='function')
+def shell(dut: DeviceAdapter) -> Shell:
+    """Return ready to use shell interface"""
+    shell = Shell(dut, timeout=20.0)
+    logger.info('Wait for prompt')
+    assert shell.wait_for_prompt()
+    return shell
+
+
+@pytest.fixture(scope='session')
+def is_mcumgr_available() -> None:
+    if not MCUmgr.is_available():
+        pytest.skip('mcumgr not available')
+
+
+@pytest.fixture()
+def mcumgr(is_mcumgr_available: None, dut: DeviceAdapter) -> Generator[MCUmgr, None, None]:
+    yield MCUmgr.create_for_serial(dut.device_config.serial)
