+++ scripts/pylib/pytest-twister-harness/src/twister_harness/plugin.py
@@ -14,8 +14,7 @@ from twister_harness.twister_harness_config import TwisterHarnessConfig
 logger = logging.getLogger(__name__)
 
 pytest_plugins = (
-    'twister_harness.fixtures.dut',
-    'twister_harness.fixtures.mcumgr'
+    'twister_harness.fixtures'
 )
